commented lines: work with them, and leave comments intact
slic3r (on my Zeus) can generate gcode with comments adjacent to
numbers, e.g.
G0 X0 Y8 F4500; Go to pre-heat location
so we have to split off the comments before assuming that every
"word" after the initial letter is part of a number.
Also, preserve unmodified any lines which are fully commented
or which do not contain movements on the axis of interest.
And use tabs for indentation (personal preference, especially for
Python, such that 1 tab = one indent-level, and visual indentation
in your editor is then just a matter of settings).